refactor: merge packages as react-components
