+++ dev/kitchen-sink/App.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import '../../packages/react-components/css/lumo/Typography.css';
 import { AppLayout } from '../../packages/react-components/src/AppLayout.js';
 import { Avatar } from '../../packages/react-components/src/Avatar.js';
-import { Board } from '../../packages/react-components-pro/src/Board.js';
+import { Board } from '../../packages/react-components/src/Board.js';
 import { Button } from '../../packages/react-components/src/Button.js';
 import { DrawerToggle } from '../../packages/react-components/src/DrawerToggle.js';
 import { Notification } from '../../packages/react-components/src/Notification.js';
