+++ dev/kitchen-sink/Row1.tsx
@@ -4,9 +4,9 @@ import { Accordion } from '../../packages/react-components/src/Accordion.js';
 import { AccordionHeading } from '../../packages/react-components/src/AccordionHeading.js';
 import { AccordionPanel } from '../../packages/react-components/src/AccordionPanel.js';
 import { AvatarGroup } from '../../packages/react-components/src/AvatarGroup.js';
-import { BoardRow } from '../../packages/react-components-pro/src/BoardRow.js';
-import { Chart } from '../../packages/react-components-pro/src/Chart.js';
-import { ChartSeries } from '../../packages/react-components-pro/src/ChartSeries.js';
+import { BoardRow } from '../../packages/react-components/src/BoardRow.js';
+import { Chart } from '../../packages/react-components/src/Chart.js';
+import { ChartSeries } from '../../packages/react-components/src/ChartSeries.js';
 
 export default function Row1() {
   return (
