+++ dev/kitchen-sink/Row9.tsx
@@ -1,4 +1,4 @@
-import { BoardRow } from '../../packages/react-components-pro/src/BoardRow.js';
+import { BoardRow } from '../../packages/react-components/src/BoardRow.js';
 import { SplitLayout } from '../../packages/react-components/src/SplitLayout.js';
 import { TextArea } from '../../packages/react-components/src/TextArea.js';
 import { Tooltip } from '../../packages/react-components/src/Tooltip.js';
