--- packages/react-components-pro/README.md
@@ -1,3 +1,0 @@
-# Vaadin React components pro
-
-React wrappers for commercial Vaadin components.
