--- packages/react-components-pro/tsconfig.build.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": ".",
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "noEmit": false,
-    "emitDeclarationOnly": true
-  },
-  "include": [
-    "./src",
-  ],
-  "exclude": [
-    "node_modules",
-    "dist"
-  ]
-}
