+++ packages/react-components/README.md
@@ -1,3 +1,3 @@
 # Vaadin React components
 
-React wrappers for Vaadin components.
+This package contains all React wrappers for Vaadin components (free and commercial).
