+++ packages/react-components/src/ChartSeries.ts
